requirements.txt
# ------------------------------------------------------------------
# Runtime dependencies.
# Pinned to compatible ranges rather than exact versions so that
# patch-level security fixes are picked up automatically.
# ------------------------------------------------------------------

# Core numerics
numpy>=1.26,<3.0
pandas>=2.1,<3.0

# Deep learning — PyTorch is the primary framework.
# TensorFlow is optional and only needed for the Keras baseline model.
torch>=2.2,<3.0
torchvision>=0.17,<1.0

# Optional: Keras baseline (install only if you need to retrain model 1).
# tensorflow>=2.15,<3.0

# Imaging
pillow>=10.0,<12.0
opencv-python-headless>=4.8,<5.0

# Metrics & ML utilities
scikit-learn>=1.4,<2.0

# Visualisation
matplotlib>=3.8,<4.0
seaborn>=0.13,<1.0

# Config / CLI / logging
pyyaml>=6.0,<7.0
rich>=13.0,<15.0
tqdm>=4.66,<5.0

# Typed config schema (lightweight alternative to Hydra)
pydantic>=2.6,<3.0
